+++ .database/WiredTiger
@@ -1,0 +1,2 @@
+WiredTiger
+WiredTiger 10.0.0: (March 18, 2020)
