+++ docker-compose.yml
@@ -1,0 +1,32 @@
+version: '3'
+
+services:
+  server:
+    build:
+      context: server/.
+      dockerfile: Dockerfile
+    container_name: accsm-server
+    volumes:
+      - server/:/usr/src/app
+    networks:
+      - accsm-network
+    ports:
+      - '8080:8080'
+    depends_on:
+      - mongo
+    env_file: .env
+    command: yarn dev
+
+  mongo:
+    image: mongo
+    container_name: accsm-mongodb
+    networks:
+      - accsm-network
+    ports:
+      - '27017:27017'
+    volumes:
+      - .database:/data/db
+
+networks:
+  accsm-network:
+    driver: bridge
