+++ server/Dockerfile
@@ -2,8 +2,6 @@
 
 FROM node:15.4.0-alpine3.12
 
-ARG NODE_ENV=development
-
 # Create App Directory
 RUN mkdir -p /usr/src/app
 WORKDIR /usr/src/app
@@ -15,8 +13,3 @@ RUN rm -rf node_modules && yarn install --frozen-lockfile
 
 # Copy app source code
 COPY . .
-
-# Exports
-EXPOSE 8080
-
-CMD ["yarn", "dev"]
