+++ server/package.json
@@ -3,14 +3,24 @@
   "version": "0.1.0",
   "description": "Dockerized server management application for Assetto Corsa Competizione built using the MERN stack.",
   "main": "server.js",
-  "repository": "https://github.com/bseth15/acc-server-manager",
+  "type": "module",
+  "scripts": {
+    "start": "node -r dotenv/config src/server.js",
+    "server": "nodemon -r dotenv/config src/server.js",
+    "client": "npm run start --prefix ../client",
+    "dev": "concurrently \"npm run server\" \"npm run client\""
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com:bseth15/acc-server-manager.git"
+  },
   "author": "Seth Brown",
   "license": "MIT",
   "private": true,
-  "scripts": {
-    "start": "node -r dotenv/config server.js",
-    "dev": "nodemon -r dotenv/config server.js"
+  "bugs": {
+    "url": "https://github.com/bseth15/acc-server-manager/issues"
   },
+  "homepage": "https://github.com/bseth15/",
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "concurrently": "^5.3.0",
