+++ server/routes/readme.txt
@@ -1,0 +1,4 @@
+api routes return json data
+
+any server rendering (ex. using template engine)
+routes should be placed in a different directory
