+++ server/server.js
@@ -2,7 +2,7 @@ const cors = require('cors');
 const express = require('express');
 const passport = require('passport');
 
-const users = require('./src/routes/api/users');
+const users = require('./routes/api/users');
 
 const port = process.env.PORT || 8080;
 const app = express();
@@ -10,16 +10,16 @@ const app = express();
 // Middleware
 app.use(cors());
 app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
+app.use(express.urlencoded({ extended: false }));
 app.use(passport.initialize());
 app.use(passport.session());
-require('./src/config/passport')(passport);
+require('./config/passport')(passport);
 
 // Use Routes
 app.use('/api/users', users);
 
 // Start Database
-require('./src/config/database');
+require('./config/database');
 
 // Start server
 app.listen(port, () => {
