--- server/src/routes/readme.txt
@@ -1,4 +1,0 @@
-api routes return json data
-
-any server rendering (ex. using template engine)
-routes should be placed in a different directory
